siakad_app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
email_validator
psycopg2-binary
argon2-cffi
python-dotenv
gunicorn
weasyprint
Flask-Limiter
Flask-Talisman
